FILES/requirements.txt
requests
shapely
mysql-connector
mysql-connector-python
flask


